changed number field types to double

+++ app/G91/ulyanov.leonid/lab4/src/main/java/ru/vsu/amm/java/entity/MoonshineData.java
@@ -6,5 +6,5 @@ import java.time.LocalDate;
 import java.util.List;
 
 public record MoonshineData(LocalDate date, String label, List<Ingredient> ingredients,
-                            float volume, float makingTime) {
+                            double volume, double makingTime) {
 }
